do not do startup check if on execute

+++ htmap/__init__.py
@@ -19,6 +19,7 @@ from typing import Tuple as _Tuple
 import logging as _logging
 import logging.handlers as _handlers
 from pathlib import Path as _Path
+import os as _os
 
 from .settings import settings, USER_SETTINGS, BASE_SETTINGS
 
@@ -27,32 +28,33 @@ _logger = _logging.getLogger(__name__)
 _logger.setLevel(_logging.DEBUG)
 _logger.addHandler(_logging.NullHandler())
 
-# ENSURE HTMAP DIR EXISTS
-_htmap_dir = _Path(settings['HTMAP_DIR'])
-if not _htmap_dir.exists():
-    try:
-        _htmap_dir.mkdir(parents = True, exist_ok = True)
-        _logger.debug(f'created HTMap dir at {_htmap_dir}')
-    except PermissionError as e:
-        raise PermissionError(f'the HTMap directory ({_htmap_dir}) needs to be writable') from e
+if _os.getenv('HTMAP_ON_EXECUTE') != '1':
+    # ENSURE HTMAP DIR EXISTS
+    _htmap_dir = _Path(settings['HTMAP_DIR'])
+    if not _htmap_dir.exists():
+        try:
+            _htmap_dir.mkdir(parents = True, exist_ok = True)
+            _logger.debug(f'created HTMap dir at {_htmap_dir}')
+        except PermissionError as e:
+            raise PermissionError(f'the HTMap directory ({_htmap_dir}) needs to be writable') from e
 
-LOG_FILE = _Path(settings['HTMAP_DIR']) / 'htmap.log'
-# SET UP LOG FILE
-_logfile_handler = _handlers.RotatingFileHandler(
-    filename = LOG_FILE,
-    mode = 'a',
-    maxBytes = 10 * 1024 * 1024,  # 10 MB
-    backupCount = 4,
-)
-_fmt = _logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-_logfile_handler.setFormatter(_fmt)
-_logfile_handler.setLevel(_logging.DEBUG)
-_logger.addHandler(_logfile_handler)
+    LOG_FILE = _Path(settings['HTMAP_DIR']) / 'htmap.log'
+    # SET UP LOG FILE
+    _logfile_handler = _handlers.RotatingFileHandler(
+        filename = LOG_FILE,
+        mode = 'a',
+        maxBytes = 10 * 1024 * 1024,  # 10 MB
+        backupCount = 4,
+    )
+    _fmt = _logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    _logfile_handler.setFormatter(_fmt)
+    _logfile_handler.setLevel(_logging.DEBUG)
+    _logger.addHandler(_logfile_handler)
 
-import shutil as _shutil
-from . import names as _names
+    import shutil as _shutil
+    from . import names as _names
 
-_shutil.rmtree(_htmap_dir / _names.REMOVED_MAPS_DIR, ignore_errors = True)
+    _shutil.rmtree(_htmap_dir / _names.REMOVED_MAPS_DIR, ignore_errors = True)
 
 from .mapping import (
     map, starmap, build_map,
